refactor: Cars 입출력문 제거 및 이름 기능 수정
1. 입출력문 제거
2. 자동차모음 생성자 파라미터 변경
