+++ src/test/java/racinggame/CarsTest.java
@@ -6,10 +6,12 @@ import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
+import racinggame.model.Cars;
+
 public class CarsTest {
 	@Test
 	void 자동차_이름_분할(){
-		Cars cars = new Cars(Arrays.asList("chang,hyuk,park".split(",")), "3");
+		Cars cars = new Cars(Arrays.asList("chang,hyuk,park".split(",")));
 		assertThat(cars.getNumberOfCars()).isEqualTo(3);
 	}
 }
